Refactor all Classic checks for icons
This hopefully fixes every possible green icon that we've got in the
UI (of which a few more were discovered while doing all this work).

Icons are now defined in a new global table, TRP3_InterfaceIcons, which
is a mapping of human-readable icon names to a list of icon file names
to be tested on startup of the addon.

When loading we check if the named icon is known to the client through
the use of the GetFileIDFromPath function and, for the first icon that
returns a valid file ID, we associate that icon name with its ID and
reference it elsewhere in the codebase.

In development builds if an icon has no valid candidate files a message
is dumped to the default error handler for remediation, and the icon
will be defaulted to the standard question mark.

All uses of icons in the codebase that I've been able to find have been
relocated to this global table giving us a single place to define and
change them in the future.