Initial updates to CI for Wago addons support
This is pending a few PRs to the upstream packager to be merged in
proper. This changeset does the following:

  - Adds the Wago project ID to our packager runs
	- Separates the workflow for a release into two, where one handles prerelease versions
	- Separates the Classic packaging step into two, where one renames the file uploaded to CF/WoWI

With these changes and the necessary PRs merged into the packager, we're
able to remove the need for our packager fork entirely and can use the
upstream one unaltered.

Note that the file renaming issue with Classic does require that we now
run the packager _three_ times; it'd be nice if we could just drop the
idea of not having "-classic" in the name for Classic project uploads
for the one site where it really matters.