Truncate characteristic fontstrings into tooltips
This modifies the fontstrings displayed in the characteristics tab
under both the "directory information" and "additional information"
sections to, if their contents are truncated, display them in a tooltip
on hover.

This largely just reuses the existing stuff in the Blizzard UI to do
so, but I've added a new TRP3_TruncatedTextTemplate which is a generic
frame that can be used in place of a fontstring to enable this behaviour
elsewhere if needed.