Fix spurious issues with Prat module (#508)
The Prat module was being registered after PLAYER_LOGIN had fired. As
Prats' module system is that of Ace3, modules are only initialized and
enabled after either ADDON_LOADED is fired or PLAYER_LOGIN occurs. By
registering our module as late as we did, it wouldn't actually be
initialized or enabled until any load-on-demand addon was subsequently
loaded and could simultaneously cause issues in the Prat code as it
might access the database for our module which wouldn't exist until
that point was reached.

To fix this and avoid having to somehow marry our _lacking_ module
system to that of Prat, I've removed our Prat module from our module
system and just registered it in the script body as intended.

Users who want to disable the Prat module can disable it from within
the Prat config UI (where all its other modules exist, so this is
consistent).

Additionally I've cleaned up all the "let's import a bunch of functions
as locals and rename half of them" in this module because quite frankly
I'm getting perturbed that functions get renamed randomly or are
ambiguous across our fifty namespaces and require a heck of a lot of
effort to actually figure out what their identities really are.