Add companion pet profile support for Classic
This initial work implements companion pet profiles for non-combat
pets and mounts. The data provided by the addon includes all possible
non-combat pets and mounts, and does not filter by the presence of their
associated items in the players' inventory.

As companion pets can't be renamed the battle pet browser doesn't
require that companions be renamed and instead keys on the name of the
item that summons the pet. This makes it harder to deal with profile
associations when dealing with multiple profiles and one pet, but this
is a limitation users will have to accept.

The resources folder has a new CompanionData script which embeds
all the data for non-combat pet companions in a way that can be
extended to support future expansions (TBC) easily.