Diet reimplementation of the nameplates module
This cuts out a lot of the features that had accumulated into the
previous implementation in favor of a leaner approach that we can
build upon based upon user requests after this launches. Consider
this to be more an MVP more than a feature-complete module.

In terms of what's implemented both Blizzard and Kui nameplates
are supported with a minimal set of options to control customizations
on a per-element basis. Visibility options are included to disable
customizations while OOC, in combat, on OOC units, or to hide nameplates
of units that have no profiles attached.

ElvUI support has been dropped for now since implementing it has
complications and I'd rather get user feedback first before committing
to any real work in that area. Glance bars are also gone, at least for
now until feedback is given by users.