Default colors to unit class if not provided
This is for consistency with things such as our tooltips where if no
custom class color is given, we use the unit class color instead.

It felt a bit weird having some nameplates have custom colors and
others not, especially in Classic, so now if a unit has no custom
class color we use the default color for their class.

This is pending user feedback and might be made a configurable.