Implement taint-free UnitPopup menu integrations (#502)
This revives the ancient PR and polishes it a bit with a few additional
features. The primary pitch of this changeset is to add right-click
options on player unitframes to open profiles and, in the case of the
local player (self), additional options to quickly change the users'
current roleplay status.

UnitPopup is a taint minefield that can be appended to via use of a
hook on UnitPopup_ShowMenu, however this comes with the caveat that
addons which do so have their entries placed below the "Cancel" or
"Close" button in each menu. This is largely unavoidable, and users
who don't want these additions can disable the module.

In addition to the work some cleanup is done elsewhere to add an
openProfile function to TRP3_API.slash, which means the hack in the
keybindings can be removed, as well as a few augmentations to the
Player model to simplify changing statuses.

Co-authored-by: Solanya <solanya@totalrp3.info>