Fix lack of sound cue when opening main window
Blizzard actually has a bunch of non-existent file IDs in the SOUNDKIT
table, so nothing was explicitly broken but the sounds that play when
opening the main window didn't work in Classic.

This is a quick fix; we'll do something similar to the interface icon
work to fix these throughout the addon some other time.