Diet reimplementation of the nameplates module (#510)
* Diet reimplementation of the nameplates module

This cuts out a lot of the features that had accumulated into the
previous implementation in favor of a leaner approach that we can
build upon based upon user requests after this launches. Consider
this to be more an MVP more than a feature-complete module.

In terms of what's implemented both Blizzard and Kui nameplates
are supported with a minimal set of options to control customizations
on a per-element basis. Visibility options are included to disable
customizations while OOC, in combat, on OOC units, or to hide nameplates
of units that have no profiles attached.

ElvUI support has been dropped for now since implementing it has
complications and I'd rather get user feedback first before committing
to any real work in that area. Glance bars are also gone, at least for
now until feedback is given by users.

* Disable old Kui nameplates module if detected

This is to avoid any conflicts that might occur with the old module
if it's loaded.

* Add warning to the hide non-roleplay units option

This feature hasn't been tested too much and I'd rather have a warning
in place explicitly that this is a volatile feature that might cause
issues with other nameplate modifications, especially to the Blizzard
nameplates.

* Comment cleanup

* Improve Blizzard logic for hiding units

Instead of unilaterally trampling on all child regions/frames we now
just selectively pick the Blizzard ones out and configure those with
hooks to record their (blizzard-set) requested state.

* Fix potential nil unit token issue

* Rework Blizzard nameplates to avoid combat taint

Customizations weren't safely applying in combat due to taint presumably
stemming from us reaching into CUF's innards with manual calls to
update health and text values.

Using the same approach for visibility, we now posthook all the
relevant setters on the name and health bar widgets which affect
text and coloring, and have an override system in place to control
if our preferred option is shown over what something else has externally
set.

* Make Blizzard nameplates cowardly

Refuse to set up the Blizzard nameplates module if we detect a few
hardcoded external addons as being enabled for the current session,
just to be sure we're not needlessly doing work in such a case or
otherwise conflicting.

* Add full title support to Blizzard nameplates

Requires that the nameplate be in name-only mode, matching the old
limitations. Felt dirty not including this.

* Change ElvUI detection to not use private table

Thank you Simpy <3

* Update help text for full title config option

Clarify the requirement for name-only mode in Kui nameplates.

* Default colors to unit class if not provided

This is for consistency with things such as our tooltips where if no
custom class color is given, we use the unit class color instead.

It felt a bit weird having some nameplates have custom colors and
others not, especially in Classic, so now if a unit has no custom
class color we use the default color for their class.

This is pending user feedback and might be made a configurable.

* Shut up bob

* Yeet DebugTools

Co-authored-by: Solanya <solanya@totalrp3.info>