Nerf hide non-roleplay units for Kui
Use the old implementation that doesn't hide the whole frame but only
the name/guild/level text widgets.

This is because there's an issue with the hide/show functionality when
a unit is initially on-screen with the "hide non-roleplay units" setting
set, then that is unticked (or customizations are disabled) and their
nameplate remains hidden.

This does mean that the setting only works while name-only mode is
enabled, since we don't touch the health bar, but this solution is
no worse than the old implementation.