Treat OOC unit nameplates as non-roleplay units
When the "Disable on OOC units" and "hide non-roleplay units" options
were combined, the behavior would be that units who were OOC would have
the customizations disabled on the nameplate but the nameplate wouldn't
be hidden as the unit was still considered a "roleplay unit" by virtue
of having a valid profile.

Now, if both options are combined all OOC units are hidden by force.