Rework visibility implementation
The visibility attribute of a nameplate is now included in the display
info table it receives from the callback as a "shouldHide" field. This
makes it so that decorators don't need to call into potentially
expensive logic to check if a nameplate should/shouldn't be visible
as this information doesn't change outside of when we supply new
display data.

With this change a lot of functions that were exposed in the Util
table are now moved as local functions to the display info driver.

An additional tweak is that the "disable customizations on OOC units"
setting has been split into a disable and hide setting separately, since
there was some confusing edge case logic and the behaviour of the
disable setting combined with "hide non-roleplay units" one was
sometimes confusing.

If "disable customizations on OOC units" is checked, you'll get the
behaviour in the current beta; OOC units will not have _any_
customizations including visibility applied to their nameplates.

If the new "hide out of character units" option is enabled then, as long
as customizations aren't disabled (either for all units or OOC ones),
the nameplate will be hidden.