Lazily activate core nameplate module
The core nameplate module is enabled all the time even if no
decorator module such as Blizzard or Kui is enabled, which needlessly
means we're doing a lot of work (such as requesting profiles and
generating display info) that'll never be shown on a nameplate.

Now, the core module has an activation step that's only triggered
when the module itself has reached the enable stage and if at least
one decorator has connected to the OnNamePlateDataUpdated callback.

In the event where nothing connects we don't set up any event listeners
for nameplate events, don't request profiles from them, and don't
do any work whatsoever.