Refresh visibility on Kui name text changes
An edge case exists where if a unit whose nameplate should be hidden
logs in within the field of view of the player, their nameplate may
not correctly be hidden and would hang around. Sometimes this would
fix itself after some time, but typically it'd require the user to
intervene by either targetting them or getting their nameplate off
and on the screen.

As a hacky workaround, when name text is updated we check if the
visibility of the widget disagrees with what our display info wants and,
if there is a mismatch, force a visibility update then and there.

As the visibililty update would call our hooked function, to avoid a
stack overflow we nop it for the duration of the update; this should
be fine but if it proves problematic then we can set a flag elsewhere
to prevent overflow. Another option would be to defer the update to
the end of the frame.