Swap to MSP-style election of nameplate modules
Our individual decorator modules now read and write to a single common
global (TRP3_NAMEPLATES_ADDON) that is either nil if no decorator addon
is provided, or a string naming the addon that will be used to provide
nameplate decorations.

The intent is that for non-Blizzard decorators they check if this is
set and, if so, don't enable themselves. If it's nil then they should
write the name of their associated addon and double-check that it hasn't
been trampled later in the load process.

For the Blizzard decorator the logic is different as this is the default
nominee when all else fails; instead it'll check if the global has been
left unset and if so assume that it's safe to install itself as king.