Buff Kui hide non-roleplay/OOC options again
Commit 4109478 made a change to nerf this mode for Kui to workaround an
issue where in the following circumstance a nameplate would incorrectly
remain hidden:

   1. Enable the "disable customizations while OOC" setting
   2. Enable the "hide non-roleplay units" setting
   3. While OOC, place camera so there is a non-roleplay unit on-screen
   4. Turn camera to get them off-screen, go IC
   5. Turn camera back, nameplate will be invisible (intended)
   6. Toggle IC/OOC state, nameplate remains invisible (bad!)

This seems to be fixed proper by just calling KuiNameplatesCore:Hide()
when hiding the nameplate, so we'll go with that and see if the issue
resurfaces.

As we're back to controlling the visibility of the nameplate proper we
no longer need the previous fix for name text showing on newly-logged-in
units, so that has been removed.