Always define sendMSPQuery function
We've got a few places in the code (nameplates, slash commands) that
assume you can send MSP requests via this function. This is only true
if the module was loaded however.

To save on duplicating checks all over the codebase, the sendMSPQuery
function is now always present irrespective of the module being enabled
but won't actually do anything unless msp_RPAddOn is set to our expected
value.