Fix health bar color conflict with Kui tank mode
The tank mode plugin like us doesn't modify the state of the nameplate
health color, so we'd keep restoring it.

We can as a workaround for now check if the health_colour_priority
field is set in the state and, if so, don't reset the health color.

Note that we might still conflict in the edge case where we decorate
the health color of a player we're in combat with, such as in a PvP
situation.