Fix Kui names when level text/health is disabled
In Kui name-only mode, the NameOnlyUpdateNameText function only changes
the name text on the fontstring if the user has either the level text
or coloring of health on the name enabled. If neither are enabled, the
text isn't updated.

Moving the SetText call out of the else block and unconditionally
replacing the contents prior to the NameOnlyUpdateNameText call
should resolve this.