Update libraries, use pkgmeta externals
So we don't need to keep updating libraries manually we'll use the
packager to fetch them at runtime like good citizens of the addon
environment.

A local copy is kept in the Libs folder so as to not break existing
checkouts, though ideally we'd have gotten rid of them there too.