Initial work for atlas support
This implements support for atlases and file IDs across the TRP UI
with a few helper templates and mixins. A few issues remain as this
was a WiP, unfortunately the main one is cursor support as these
support neither file IDs nor atlases.

This work will likely be revisited much later in the future if/when
a solution for these is found, but the technical complications these
changes would bring for the sake of ~50 additional icons is likely
not worth it.