Route logging to _DebugLog if installed
This removes all the log spam going to the chat frame on initial
login with a dev build, as well as the need to set up a "Logs" window,
by routing all log messages to the _DebugLog addon if it's loaded on
the client.

Logs can be accessed via `/dl` and are categorised by the name of the
logger that sent them. The tool provides filtering, search, and export
capabilities which I think was the initial aim for the logging system
anyway.