Fix mature tooltips overlapping GameTooltip
Fixes #541. When a profile with mature content is assigned to the
mouseover unit, with the "Hide original tooltip" setting checked
we'd both fail to hide said tooltip and would overlap our mature
content warning right over it.

This changes the logic to treat the mature contnet warning in the same
way as the ignored profile one, and should anchor it such that it sits
atop GameTooltip.