Protect all text inside square brackets
Square brackets are commonly used as we do for addons that want to
transmit links across the chat system and reconstitute them on the
other end; one such example would be WeakAuras which uses the format
"[WeakAuras: Character - Aura Name]".

Previously if any such texts contained speech, emote, or OOC triggers
they'd be mangled like our own. In the interest of cross-addon
compatibility this change now makes it so that any text housed inside
square brackets is no longer a candidate for these color replacements.

It should be fairly rare for someone to use square brackets and inside
them place speech/emote/OOC markers; if someone _does_ complain we can
consider a change to instead just replace "[Word: ...]" style texts;
this isn't done now however because not all links necessarily follow
the above format (with the colon) as rigidly as we may hope. Basically
we'll wait and see if anyone complains first, since this implementation
is a simpler pattern and more generally useful.