Hide trimmed Currently/OOC tooltip texts if empty
If a profile is displayed in the tooltip contains a CU/CO field with
whitespace then the resulting tooltip would show the respective header
but have no content beneath it, as trimming of the field occurs after
the empty check.

This changeset makes us additionally check if the trimmed text for these
fields is empty and, if so, prevents us from inserting empty sections
into the tooltip for these fields.