Add customization of link colors in descriptions
If a link that gets transformed to HTML is preceeded by an immediate
"{col}" tag we no longer insert color markup for the link and instead
allow that tag to control the color of the link.

A couple of pattern changes were also made for the matching of
"{link}" style tags to resolve a few issues, these might be breaking
changes and - if so - aren't intentional. Testing would be needed
here.