Update power type global strings (#547)
Patch 9.1 seems to be removing the MAELSTROM global string entirely
which breaks us for that patch. There exists an equivalent in the
form of POWER_TYPE_MAELSTROM, which exists on all clients so we'll
use that instead.

For all other power types that have corresponding POWER_TYPE_* global
string constants these have also been changed, in case Blizzard makes
more changes of this kind in the future.