Protect "[TRP3:...]" links from chat transforms (#544)
* Protect "[TRP3:...]" links from chat transforms

Fixes #540. Previously a chat link that contained any OOC, speech or
emote pattern matches would have their inner text mangled and replaced
with the appropriate color sequences, as it was possible for the
message handler that processes these to be installed after the one
used for text color processing.

This rewrites the message protection code; the old approach would first
find the patterns it was interested in and replace them with a temporary
pre/postfix marker, it'd then check the things it didn't want to mess
up and remove those markers, and then replace the remaining markers with
the intended color sequence.

Instead, this new approach changes the order of operations - we find
things that should be protected (hyperlinks, declensions, and our own
textual link format) and replace those with a uniquely generated escape
sequences, borrowing from Blizzards' "|K" protected strings, and store
the replacements in a table. Message processing then occurs as normal
after that with the protected strings stashed away, and then they're
restored as the last step of the chat handler. This approach ends up
being safer and more maintainable, as well as more efficient as we
need less closures and string replacements.

* Protect all text inside square brackets

Square brackets are commonly used as we do for addons that want to
transmit links across the chat system and reconstitute them on the
other end; one such example would be WeakAuras which uses the format
"[WeakAuras: Character - Aura Name]".

Previously if any such texts contained speech, emote, or OOC triggers
they'd be mangled like our own. In the interest of cross-addon
compatibility this change now makes it so that any text housed inside
square brackets is no longer a candidate for these color replacements.

It should be fairly rare for someone to use square brackets and inside
them place speech/emote/OOC markers; if someone _does_ complain we can
consider a change to instead just replace "[Word: ...]" style texts;
this isn't done now however because not all links necessarily follow
the above format (with the colon) as rigidly as we may hope. Basically
we'll wait and see if anyone complains first, since this implementation
is a simpler pattern and more generally useful.