Split profile information responses into queues (#555)
Currently, when responding to a request for profile information we
respond without making use of the queue feature of Chomp and place
all messages into the default queue ("prefix-channel-target"). As
when sending such requests we ask for the About text second, this
causes that field to essentially block the sending back of all other
data until the About text has been fully received.

This change makes it so that individual information types are placed
into their own queue of the form "prefix-subprefix-type-target"; this
means that any information sent back in response to an incoming request
no longer blocks other pieces of data being sent to that player.

For normal profiles which are expected to have larger about texts than
any other section, the expectation is that this should make those feel
a bit snappier as about text would no longer block the misc/character
tables. For profiles that are just generally huge all around, the
difference shouldn't hopefully be noticeably slower.