Recursively unprotect string content
The hypothetical case is that a string like "[|Hlinkdata:1:2:3|hLink|H]"
would be protected twice as the patterns applied to protect contents
would first match the inner "|H" string and then the outer brackets.

On restoring strings we'd only do one loop and restore a string like
"[|Ktrp1|k]", so we need to continually unprotect strings until we
stop making replacements.