Clean up version check logic
This removes the version check logic that was left in a not-fully
working state and isn't really applicable on 9.x clients going forward
anyway, so probably not worth resuscitating.

From what I can tell the original history of this change was intended to
generate a newly named "__version.lua" file for each packaged release,
such that when the client loaded the addon it'd fail to find the newly
named file and we'd bail early due to the version metadata being unset.

This hasn't been the case for a while however - we don't generate a
named "__version.lua" script specific to the release so the version
metadata is always present.

Further, with 9.x clients onward new files are detected on UI reload so
the approach of relying on the client to not load the new metadata file
won't work.

Finally, the version check logic itself assumed that setting TRP3_API
to nil would break the rest of the addon, but it didn't actually do
anything because it was assigning nil to a local variable instead - and
most other parts of the addon don't access TRP3_API through a global
anyway instead opting to use the varargs available in the script body.

Due to all the above I think it's probably best that we just remove this
stuff since it's not in a working state as-is and the implementation of
it fundamentally can't really work without larger changes to the addon,
which for Retail isn't that necessary due to the UI reload changes for
new files anyway.