Rework addon credit texts
The credits string in our localization is the sole reason why the
individual Classic and Retail projects need to currently exist, this
changeset refactors the generation of the credits in the dashboard to
be more piecemeal and avoid hardcoded textures, links, and other things
that have fallen out of sync between the projects.

No change to the structure of the credits has been made, nor have any
significant adjustments been made to roles other than to give Sol the
"Project Manager" role and to rename "Contributor" to "Developer". All
people who were on the credits before are still there now, but the
layout of the acknowledgements has changed slightly to remove some
hardcoded references.