Add BCC client checks where needed
This is largely untested at present, thankfully with the icons
work already done the number of places that need tweaks is reduced.

The "wrong client" check has been removed to keep things simple and
since our long-term goal is probably to take advantage of the single
package things where possible. Other than that it's just a case of
adding an is_bcc constant and sprinkling that around to all the
existing places where these things are tested.